refactor: domain 패키지추가
