+++ src/main/java/lotto/domain/NumberMatcher.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public interface NumberMatcher {
